pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "django-mfa2>=3.1.1",
    "django-localflavor>=5.0",
    "django-crispy-forms>=2.4",
    "crispy-bootstrap5>=2025.6",
    "psycopg2-binary>=2.9.10",
    "dj-database-url>=3.0.1",
    "whitenoise>=6.11.0",
    "gunicorn>=23.0.0",
    "uvicorn>=0.37.0"
]
maintainers = [
  {name = "Cristian Berceanu", email = "cristi92b@gmail.com"}
]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"